fix various windows compilation warnings (2)

+++ src/var.cpp
@@ -1051,7 +1051,7 @@ uint32_t jitc_var_resize(uint32_t index, size_t size) {
         v2.type = v->type;
         v2.backend = v->backend;
         v2.placeholder = v->placeholder;
-        v2.size = size;
+        v2.size = (uint32_t) size;
         v2.dep[0] = index;
         v2.stmt = (char *) (((JitBackend) v->backend == JitBackend::CUDA)
                             ? "mov.$t0 $r0, $r1"
